# blog-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|f1j0D8H7gVxW9abc..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход пользователя из системы",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован (нет токена)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Создать комментарий",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "commentable_type",
                                    "commentable_id"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Отличный пост!"
                                    },
                                    "commentable_type": {
                                        "type": "string",
                                        "example": "App\\\\Models\\\\Post"
                                    },
                                    "commentable_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий создан"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Обновить комментарий",
                "operationId": "39e62963cbf897a1b2168683878e0a46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Обновленный комментарий"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий обновлён"
                    },
                    "403": {
                        "description": "Нет прав на редактирование"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удалить комментарий",
                "operationId": "721e5dc0c32ef81276203df965767606",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Комментарий удалён"
                    },
                    "403": {
                        "description": "Нет прав на удаление"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Комментарии текущего пользователя",
                "operationId": "6c8f7818692fdbe5c10bfa480558977b",
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{userId}/comments/active-posts": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Комментарии пользователя к активным постам",
                "operationId": "387408994a244bc8cffe7467d6f966fd",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        },
        "/api/posts/{postId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Комментарии поста",
                "operationId": "3b0d59357cbbe36767df85113dcc823c",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID поста",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        },
        "/api/comments/{commentId}/replies": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Ответы на комментарий",
                "operationId": "579ec81736ac0e01c0b8694bbbf4925b",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        },
...
    },
...
}